refactor(settings): extract record codec
